+++ Backend/erpc/src/main/java/com/cineverse/erpc/salesopp/opportunity/repository/SalesOppDeleteRequestRepository.java
@@ -3,5 +3,5 @@ package com.cineverse.erpc.salesopp.opportunity.repository;
 import com.cineverse.erpc.salesopp.opportunity.aggregate.SalesOppDeleteRequest;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface SalesOppDeleteRequestRepository extends JpaRepository<SalesOppDeleteRequest, Integer> {
+public interface SalesOppDeleteRequestRepository extends JpaRepository<SalesOppDeleteRequest, Long> {
 }
